fix async middleware
